Fix to features to fix problem breaking autorelease.
Currently plugin2oc is using base-features... this is an
old (no longer used) name for what became
features-base.

Because features-base is no longer in controller, when we build
non-SNAPSHOT features in autorelease, this breaks plugin2oc.

The reason this was not seen outside of autorelease, is because
there are still ancient SNAPSHOT artifacts for base-features
in nexus.

Change-Id: I3eacbf464c9fd4899c72b9bb23e1e610a75944a8
Signed-off-by: Ed Warnicke <eaw@cisco.com>